Changed penalty constants int -> double

+++ src/main/java/hr/fer/seminar/solution/SolutionEVRP.java
@@ -17,11 +17,11 @@ import io.jenetics.Genotype;
 
 public abstract class SolutionEVRP<T extends Gene<?, T>> {
 
-	private static final int VEHICLE_PENALTY_CONSTANT = 0;
+	private static final double VEHICLE_PENALTY_CONSTANT = 0;
 
-	private static final int ENERGY_PENALTY_CONSTANT = 1;
+	private static final double ENERGY_PENALTY_CONSTANT = 1;
 	
-	private static final int LATENCY_PENALTY_CONSTANT = 0;
+	private static final double LATENCY_PENALTY_CONSTANT = 0;
 
 	protected final EVRPProblem problem;
